Update the vim and macvim homebrew formulae
The formulae don't accept options (see [Homebrew/homebrew-core#34411][1]
and [Homebrew/homebrew-core#34421][2]).

Also, the homebrew-provided macvim is buggy on macOS Mojave and building
MacVim from HEAD fails since it tries to use the system ruby (and that
can't be customized since there are no more options).

For that reason, I use the "macvim" cask which works fine.

[1]: https://github.com/Homebrew/homebrew-core/pull/34411
[2]: https://github.com/Homebrew/homebrew-core/pull/34421